changed long long to just int

+++ cpp02/ex00/Fixed.hpp
@@ -6,7 +6,7 @@
 
 class Fixed {
 	private:
-		long long int		_fixedPointValue;
+		int		_fixedPointValue;
 		static const int	_fractionalBits;
 
 	public:
